+++ app/src/main/java/ch/ethz/inf/vs/fstreun/payapp/GroupActivity.java
@@ -37,8 +37,11 @@ import java.util.UUID;
 
 import ch.ethz.inf.vs.fstreun.finance.Group;
 import ch.ethz.inf.vs.fstreun.finance.Participant;
+import ch.ethz.inf.vs.fstreun.finance.SimpleGroup;
 import ch.ethz.inf.vs.fstreun.finance.Transaction;
 import ch.ethz.inf.vs.fstreun.network.DataSyncSubscribeService;
+import ch.ethz.inf.vs.fstreun.payapp.ListAdapters.ListParticipantsAdapter;
+import ch.ethz.inf.vs.fstreun.payapp.filemanager.DataService;
 import ch.ethz.inf.vs.fstreun.payapp.filemanager.FileHelper;
 
 public class GroupActivity extends AppCompatActivity implements DataSyncSubscribeService.DataSyncCallback{
