+++ app/src/main/java/ch/ethz/inf/vs/fstreun/payapp/ListAdapters/ListSimpleNameAdapter.java
@@ -1,4 +1,4 @@
-package ch.ethz.inf.vs.fstreun.payapp;
+package ch.ethz.inf.vs.fstreun.payapp.ListAdapters;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -11,32 +11,32 @@ import android.widget.TextView;
 
 import java.util.List;
 
+import ch.ethz.inf.vs.fstreun.payapp.R;
+
 /**
- * Created by fabio on 11/24/17.
+ * Created by fabio on 12/6/17.
  */
 
-public class ListGroupAdapter extends ArrayAdapter<SimpleGroup> {
-
-    public ListGroupAdapter(Context context, List<SimpleGroup> content){
-        super(context, R.layout.list_item_group, content);
+public class ListSimpleNameAdapter extends ArrayAdapter<String> {
+    public ListSimpleNameAdapter(@NonNull Context context, @NonNull List<String> objects) {
+        super(context, R.layout.list_item_simple_name, objects);
     }
 
     @NonNull
     @Override
     public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
-
         // Get the data item for this position
-        String groupName = getItem(position).groupName;
+        final String string = getItem(position);
         // Check if an existing view is being reused, otherwise inflate the view
         if (convertView == null) {
-            convertView = LayoutInflater.from(getContext()).inflate(R.layout.list_item_group, parent, false);
+            convertView = LayoutInflater.from(getContext()).inflate(R.layout.list_item_simple_name, parent, false);
         }
+
         // Lookup view for data population
-        TextView tvName = (TextView) convertView.findViewById(R.id.textView_groupName);
-        // Populate the data into the template view using the data object
-        tvName.setText(groupName);
+        TextView tvMain = convertView.findViewById(R.id.textView_main);
+
+        tvMain.setText(string);
         // Return the completed view to render on screen
         return convertView;
-
     }
 }
